# SpringAlbumsAPI/api-docs.json
{"openapi":"3.1.0","info":{"title":"User API","description":"Spring Boot RESTful Albums API","contact":{"name":"Mandar Dhamale","url":"mandardhamale.com ","email":"mandardhamale@gmail.com"},"version":"Verstion 1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Auth Controller","description":"Controller for account management"},{"name":"Home Controller","description":"Controller for home management"},{"name":"Album Controller","description":"Controller for album & photo management"}],"paths":{"/api/v1/auth/users/{user_id}/update-authorities":{"put":{"tags":["Auth Controller"],"summary":"Update authorities","operationId":"updateAuthorities","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthoritiesDTO"}}},"required":true},"responses":{"200":{"description":"Update authorities of users (admin only)","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"401":{"description":"Please check access token","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"403":{"description":"Scope restriction","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"404":{"description":"User not found","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}},"security":[{"mrd-api":[]}]}},"/api/v1/auth/profile/update-password":{"put":{"tags":["Auth Controller"],"summary":"Update password","operationId":"updatePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordDTO"}}},"required":true},"responses":{"200":{"description":"Password updated","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"401":{"description":"Please check access token","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"403":{"description":"Scope restriction","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}},"security":[{"mrd-api":[]}]}},"/api/v1/album/albums/{album_id}/update":{"put":{"tags":["Album Controller"],"summary":"Update album","operationId":"updateAlbum","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumPayloadDTO"}}},"required":true},"responses":{"400":{"description":"Please add a valid description","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"204":{"description":"Album updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"security":[{"mrd-api":[]}]}},"/api/v1/album/albums/{album_id}/photos/{photo_id}/update":{"put":{"tags":["Album Controller"],"summary":"Update photo","operationId":"updatePhoto","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoPayloadDTO"}}},"required":true},"responses":{"400":{"description":"Please add a valid description","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoViewDTO"}}}},"204":{"description":"Photo updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoViewDTO"}}}},"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoViewDTO"}}}}},"security":[{"mrd-api":[]}]}},"/api/v1/auth/users/add":{"post":{"tags":["Auth Controller"],"summary":"Add a new user","operationId":"addUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountDTO"}}},"required":true},"responses":{"400":{"description":"Please enter a valid email and password length between 6 & 20","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"201":{"description":"Account added","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"Account added","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/api/v1/auth/token":{"post":{"tags":["Auth Controller"],"summary":"Generate new token","operationId":"token","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDTO"}}},"required":true},"responses":{"400":{"description":"Authentication error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenDTO"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenDTO"}}}}}}},"/api/v1/album/{album_id}/upload-photos":{"post":{"tags":["Album Controller"],"summary":"Upload photos in album","operationId":"addPhotos","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"files":{"type":"array","items":{"type":"string","format":"binary"}}},"required":["files"]}}}},"responses":{"400":{"description":"Please check the payload or token","content":{"*/*":{"schema":{"type":"array","items":{"type":"object","additionalProperties":{"type":"array","items":{}}}}}}},"401":{"description":"Invalid token error","content":{"*/*":{"schema":{"type":"array","items":{"type":"object","additionalProperties":{"type":"array","items":{}}}}}}}},"security":[{"mrd-api":[]}]}},"/api/v1/album/add":{"post":{"tags":["Album Controller"],"summary":"Add new album","operationId":"addAlbum","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumPayloadDTO"}}},"required":true},"responses":{"200":{"description":"Album created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"400":{"description":"Please add a valid description","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"security":[{"mrd-api":[]}]}},"/api/v1/auth/users":{"get":{"tags":["Auth Controller"],"summary":"List all users","operationId":"getAllAccounts","responses":{"200":{"description":"List of users","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountViewDTO"}}}}},"401":{"description":"Please check access token","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountViewDTO"}}}}}},"security":[{"mrd-api":[]}]}},"/api/v1/auth/profile":{"get":{"tags":["Auth Controller"],"summary":"View profile","operationId":"getProfile","responses":{"200":{"description":"List of users","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileDTO"}}}},"401":{"description":"Please check access token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileDTO"}}}},"403":{"description":"Scope restriction","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileDTO"}}}}},"security":[{"mrd-api":[]}]}},"/api/v1/album/albums":{"get":{"tags":["Album Controller"],"summary":"List album","operationId":"albums","responses":{"200":{"description":"List of albums","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"401":{"description":"Token missing","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"403":{"description":"Token error","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}}},"security":[{"mrd-api":[]}]}},"/api/v1/album/albums/{album_id}":{"get":{"tags":["Album Controller"],"summary":"List album by album id","operationId":"albumsByAlbumId","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"List of albums by album id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"401":{"description":"Token missing","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"403":{"description":"Token error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"security":[{"mrd-api":[]}]}},"/api/v1/album/albums/{album_id}/photos/{photo_id}/download-thumbnail":{"get":{"tags":["Album Controller"],"operationId":"downloadThumbnail","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"mrd-api":[]}]}},"/api/v1/album/albums/{album_id}/photos/{photo_id}/download-photo":{"get":{"tags":["Album Controller"],"operationId":"downloadPhoto","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"mrd-api":[]}]}},"/":{"get":{"tags":["Home Controller"],"summary":"Home page","operationId":"home","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/auth/profile/delete":{"delete":{"tags":["Auth Controller"],"summary":"Delete profile","operationId":"deleteProfile","responses":{"200":{"description":"Profile deleted","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"401":{"description":"Please check access token","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"403":{"description":"Scope restriction","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}},"security":[{"mrd-api":[]}]}},"/api/v1/album/albums/{album_id}/photos/{photo_id}/delete":{"delete":{"tags":["Album Controller"],"summary":"Delete a photo","operationId":"deletePhoto","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"202":{"description":"Photo delete","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}},"security":[{"mrd-api":[]}]}},"/api/v1/album/albums/{album_id}/delete":{"delete":{"tags":["Album Controller"],"summary":"Delete a photo","operationId":"deleteAlbum","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"202":{"description":"Album deleted","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}},"security":[{"mrd-api":[]}]}}},"components":{"schemas":{"AuthoritiesDTO":{"type":"object","properties":{"authorities":{"type":"string","description":"Authorities (add multiple authorities separated with space","example":"USER","minLength":1}},"required":["authorities"]},"PasswordDTO":{"type":"object","properties":{"password":{"type":"string","description":"Password","example":"hj37e!","maxLength":20,"minLength":6}},"required":["password"]},"AlbumPayloadDTO":{"type":"object","properties":{"name":{"type":"string","description":"Album name","example":"Travel","minLength":1},"description":{"type":"string","description":"Album description","example":"England tour photos","minLength":1}},"required":["description","name"]},"AlbumViewDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string","description":"Album name","example":"Travel","minLength":1},"description":{"type":"string","description":"Album description","example":"England tour photos","minLength":1},"photos":{"type":"array","items":{"$ref":"#/components/schemas/PhotoDTO"}}},"required":["description","name"]},"PhotoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"fileName":{"type":"string"},"downloadLink":{"type":"string"}}},"PhotoPayloadDTO":{"type":"object","properties":{"name":{"type":"string","description":"Photo name","example":"Selfie","minLength":1},"description":{"type":"string","description":"Description of the photo","minLength":1}},"required":["description","name"]},"PhotoViewDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"}}},"AccountDTO":{"type":"object","properties":{"email":{"type":"string","description":"Email address","example":"abc@gmail.com"},"password":{"type":"string","description":"Password","example":"hj37e!","maxLength":20,"minLength":6},"authorities":{"type":"string","description":"Role","example":"USER ADMIN"}},"required":["email","password"]},"UserLoginDTO":{"type":"object","properties":{"email":{"type":"string","description":"Email address","example":"mandar@gmail.com"},"password":{"type":"string","description":"Password","example":"password@1234","maxLength":20,"minLength":6}},"required":["email","password"]},"TokenDTO":{"type":"object","properties":{"token":{"type":"string"}}},"AccountViewDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"authorities":{"type":"string"}}},"ProfileDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"authorities":{"type":"string"}}}},"securitySchemes":{"mrd-api":{"type":"http","in":"header","scheme":"bearer"}}}}